+++ clang/docs/ReleaseNotes.rst
@@ -262,6 +262,7 @@ Miscellaneous Clang Crashes Fixed
 - Fixed a crash when using loop hint with a value dependent argument inside a
   generic lambda. (#GH172289)
 - Fixed a crash in C++ overload resolution with ``_Atomic``-qualified argument types. (#GH170433)
+- Fixed a crash in ``_Generic`` selection when an association type is an error type. (#GH176929)
 
 OpenACC Specific Changes
 ------------------------